requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.3
colorama==0.4.6
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
filelock==3.20.0
fsspec==2025.9.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
PyYAML==6.0.3
regex==2025.10.23
requests==2.31.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.15.0
tzdata==2025.2
unicorn==2.1.4
urllib3==2.5.0
